+++ test/integration/gen_class/argparser_arg_dir_chk_crt.py
@@ -34,6 +34,9 @@ import version
 
 __version__ = version.__version__
 
+# Global
+PERM4 = "444"
+
 
 class UnitTest(unittest.TestCase):
 
@@ -65,17 +68,19 @@ class UnitTest(unittest.TestCase):
 
         """
 
+        p_name = "program.py"
+
         self.base_path = "test/integration/gen_class/tmp"
         self.dir3 = os.path.join(self.base_path, "dir3")
         self.dir4 = os.path.join(self.base_path, "dir4")
         self.dir5 = os.path.join(self.base_path, "dir5")
         self.dir6 = os.path.join(self.base_path, "dir6")
 
-        self.argv = ["program.py"]
-        self.argv2 = ["program.py", "-d", self.dir3]
-        self.argv3 = ["program.py", "-d", self.dir3, "-g", self.dir4]
-        self.argv4 = ["program.py", "-d", self.dir5]
-        self.argv5 = ["program.py", "-d", self.dir6]
+        self.argv = [p_name]
+        self.argv2 = [p_name, "-d", self.dir3]
+        self.argv3 = [p_name, "-d", self.dir3, "-g", self.dir4]
+        self.argv4 = [p_name, "-d", self.dir5]
+        self.argv5 = [p_name, "-d", self.dir6]
 
         self.opt_val = ["-d", "-g"]
 
@@ -149,8 +154,10 @@ class UnitTest(unittest.TestCase):
 
         """
 
+        global PERM4
+
         os.mkdir(self.dir3)
-        os.chmod(self.dir3, 0o444)
+        os.chmod(self.dir3, int(PERM4, 8))
         args_array = gen_class.ArgParser(
             self.argv2, opt_val=self.opt_val, dir_chk=self.dir_chk2,
             do_parse=True)