+++ test/unit/gen_class/argparser_arg_dir_crt.py
@@ -68,10 +68,11 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        self.argv = ["program.py"]
-        self.argv2 = ["program.py", "-d", "/path/dir1"]
-        self.argv3 = [
-            "program.py", "-d", "/path/dir1", "-g", "/path/dir2"]
+        p_name = "program.py"
+
+        self.argv = [p_name]
+        self.argv2 = [p_name, "-d", "/path/dir1"]
+        self.argv3 = [p_name, "-d", "/path/dir1", "-g", "/path/dir2"]
         self.opt_val = ["-d", "-g"]
         self.dir_perms_crt = {}
         self.dir_perms_crt2 = {"-d": 1}